4th version - wrap low-level connection in SimpleConnection (and use SimpleConnectionMaker to create connections) to make testing easier
