+++ client/src/main/java/com/networknt/client/http/SimpleConnection.java
@@ -1,0 +1,9 @@
+package com.networknt.client.http;
+
+public interface SimpleConnection<Connection> {
+    public boolean isOpen();
+    public Connection getRawConnection();
+    public boolean isMultiplexingSupported();
+    public String getLocalAddress();
+    public void safeClose();
+}
