+++ client/src/main/java/com/networknt/client/http/SimpleConnectionPool.java
@@ -10,21 +10,23 @@ import java.util.concurrent.ConcurrentHashMap;
  */
 public class SimpleConnectionPool {
     private Map<URI, SimpleURIConnectionPool> pools = new ConcurrentHashMap<>();
+    private SimpleConnectionMaker connectionMaker;
     private long expireTime;
     private int poolSize;
 
     private SimpleConnectionPool() {}
 
-    public SimpleConnectionPool(long expireTime, int poolSize) {
+    public SimpleConnectionPool(long expireTime, int poolSize, SimpleConnectionMaker connectionMaker) {
         this.expireTime = expireTime;
         this.poolSize = poolSize;
+        this.connectionMaker = connectionMaker;
     }
 
     public SimpleConnectionHolder.ConnectionToken borrow(long createConnectionTimeout, boolean isHttp2, URI uri) {
         if(!pools.containsKey(uri)) {
             synchronized (pools) {
                 if (!pools.containsKey(uri))
-                    pools.put(uri, new SimpleURIConnectionPool(uri, expireTime, poolSize));
+                    pools.put(uri, new SimpleURIConnectionPool(uri, expireTime, poolSize, connectionMaker));
             }
         }
         return pools.get(uri).borrow(createConnectionTimeout, isHttp2);
